dao buds fixed

+++ time_tracker_web_app/src/main/java/com/timetracker/timet_racker_web_app/dao/repository/DateNoteRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 import java.sql.Timestamp;
 import java.util.List;
 
-public interface DateNoteRepository extends CrudRepository<DateNote, Long> {
+public interface DateNoteRepository extends CrudRepository<DateNote, Integer> {
     public List<DateNote> getAllByUsersByUser(User user);
     public List<DateNote> getAllByUsersByUserAndDate(User user, Timestamp date);
 }
